improve debian8 support

+++ setupenv.sh
@@ -7,7 +7,7 @@ trap 'err_report $LINENO $FILENAME $RUNNINGSCRIPT; exit 1' ERR
 
 #OpenRTM-aist
 sudo apt-get -y install autoconf
-if [ "$DIST_VER" = "16.04" ]; then
+if [ "$DIST_VER" = "16.04" ] || [ "$DIST_VER" = "8" ]; then
     sudo apt-get -y install libtool-bin
 else
     sudo apt-get -y install libtool
@@ -33,11 +33,19 @@ if [ "$INTERNAL_MACHINE" -eq 0 ]; then
 	sudo apt-get -y install liboctomap-dev
     fi
 fi
+
 #hrpsys-base
 sudo apt-get -y --force-yes install libxml2-dev libsdl-dev libglew-dev libopencv-dev libcvaux-dev libhighgui-dev libqhull-dev freeglut3-dev libxmu-dev python-dev libboost-python-dev ipython openrtm-aist-python
+
 #hmc2
 sudo apt-get -y install libyaml-dev libncurses5-dev
 
+#state-observation
+if [ "$DIST_VER" = "8" ]; then
+    sudo apt-get install libboost-test-dev
+fi
+
+#savedbg
 if [ "$ENABLE_SAVEDBG" -eq 1 ]; then
     if [ "$DIST_VER" = 14.04 ]; then
         REALPATH=realpath
@@ -48,42 +56,45 @@ if [ "$ENABLE_SAVEDBG" -eq 1 ]; then
 fi
 
 if [ "$INTERNAL_MACHINE" -eq 0 ]; then
-cd $SRC_DIR/choreonoid/misc/script
-./install-requisites-$DIST_KIND-$DIST_VER.sh
+    #choreonoid
+    cd $SRC_DIR/choreonoid/misc/script
+    ./install-requisites-$DIST_KIND-$DIST_VER.sh
 
-#hrpcnoid
-sudo apt-get -y install libzbar-dev python-matplotlib
+    #hrpcnoid
+    sudo apt-get -y install libzbar-dev python-matplotlib
 else
-cd $WORKSPACE
-sudo rm -fr 3.2.5.tar.gz eigen-eigen-bdd17ee3b1b3
-wget -q http://bitbucket.org/eigen/eigen/get/3.2.5.tar.gz
-tar zxvf 3.2.5.tar.gz
-cd eigen-eigen-bdd17ee3b1b3
-mkdir build
-cd build
-cmake -DCMAKE_INSTALL_PREFIX=/usr/local ..
-sudo make -j$MAKE_THREADS_NUMBER install
-cd ../..
-sudo rm -fr 3.2.5.tar.gz eigen-eigen-bdd17ee3b1b3
-
-sudo apt-get -y install libbz2-dev
-rm -f boost_1_54_0.tar.gz
-wget -q http://sourceforge.net/projects/boost/files/boost/1.54.0/boost_1_54_0.tar.gz
-tar zxvf boost_1_54_0.tar.gz
-cd boost_1_54_0
-./bootstrap.sh
-sudo ./b2 install -j2 --prefix=/usr/local
-cd ..
-sudo rm -fr boost_1_54_0.tar.gz boost_1_54_0
-
-wget -q http://sourceforge.net/projects/collada-dom/files/Collada%20DOM/Collada%20DOM%202.4/collada-dom-2.4.0.tgz
-tar zxvf collada-dom-2.4.0.tgz
-cd collada-dom-2.4.0
-mkdir build
-cd build
-cmake -DCMAKE_INSTALL_PREFIX=/usr/local ..
-sudo make -j$MAKE_THREADS_NUMBER install
-cd ../..
-sudo rm -fr collada-dom-2.4.0.tgz collada-dom-2.4.0
+    if [ "$DIST_VER" = 7 ]; then
+	cd $WORKSPACE
+	sudo rm -fr 3.2.5.tar.gz eigen-eigen-bdd17ee3b1b3
+	wget -q http://bitbucket.org/eigen/eigen/get/3.2.5.tar.gz
+	tar zxvf 3.2.5.tar.gz
+	cd eigen-eigen-bdd17ee3b1b3
+	mkdir build
+	cd build
+	cmake -DCMAKE_INSTALL_PREFIX=/usr/local ..
+	sudo make -j$MAKE_THREADS_NUMBER install
+	cd ../..
+	sudo rm -fr 3.2.5.tar.gz eigen-eigen-bdd17ee3b1b3
+	
+	sudo apt-get -y install libbz2-dev
+	rm -f boost_1_54_0.tar.gz
+	wget -q http://sourceforge.net/projects/boost/files/boost/1.54.0/boost_1_54_0.tar.gz
+	tar zxvf boost_1_54_0.tar.gz
+	cd boost_1_54_0
+	./bootstrap.sh
+	sudo ./b2 install -j2 --prefix=/usr/local
+	cd ..
+	sudo rm -fr boost_1_54_0.tar.gz boost_1_54_0
+	
+	wget -q http://sourceforge.net/projects/collada-dom/files/Collada%20DOM/Collada%20DOM%202.4/collada-dom-2.4.0.tgz
+	tar zxvf collada-dom-2.4.0.tgz
+	cd collada-dom-2.4.0
+	mkdir build
+	cd build
+	cmake -DCMAKE_INSTALL_PREFIX=/usr/local ..
+	sudo make -j$MAKE_THREADS_NUMBER install
+	cd ../..
+	sudo rm -fr collada-dom-2.4.0.tgz collada-dom-2.4.0
+    fi
 fi
